Partially revert 0033224c25a
font_driver_get_message_width() can return -1 if it fails,
ozone_cache_footer_label() has a fallback for this case

+++ menu/drivers/ozone.c
@@ -336,7 +336,7 @@ typedef struct
 {
    const char *str;
    float x;
-   size_t width;
+   int width; /* -1 on error */
    bool show;
 } ozone_footer_label_t;
